+++ rng/gsl_rng.h
@@ -7,7 +7,7 @@ typedef struct {
   unsigned long int max ;
   size_t size ;
   void * state ;
-  void (* set)(void * state, unsigned int seed) ;
+  void (* set)(void * state, unsigned long int seed) ;
   unsigned long int (* get)(void * state) ;
 } gsl_rng ;
 
@@ -15,7 +15,7 @@ typedef struct {
   const char * name ;
   unsigned long int max ;
   size_t size ;
-  void (* set)(void * state, unsigned int seed) ;
+  void (* set)(void * state, unsigned long int seed) ;
   unsigned long int (* get)(void * state) ;
 } gsl_rng_type ;
 
@@ -33,7 +33,7 @@ extern const gsl_rng_type * gsl_rng_uni ;
 extern const gsl_rng_type * gsl_rng_uni32 ;
 extern const gsl_rng_type * gsl_rng_zuf ;
 
-extern unsigned int gsl_rng_default_seed ;
+extern unsigned long int gsl_rng_default_seed ;
 
 unsigned long int gsl_rng_get (const gsl_rng * r) ;
 double gsl_rng_get_uni (const gsl_rng * r) ;
@@ -42,7 +42,7 @@ gsl_rng * gsl_rng_alloc (const gsl_rng_type * T) ;
 gsl_rng * gsl_rng_clone (const gsl_rng * r) ;
 void gsl_rng_free (gsl_rng * r) ;
 
-void gsl_rng_set (const gsl_rng * r, unsigned int seed) ;
+void gsl_rng_set (const gsl_rng * r, unsigned long int seed) ;
 unsigned long int gsl_rng_max (const gsl_rng * r) ;
 const char * gsl_rng_name (const gsl_rng * r) ;
 void gsl_rng_print_state (const gsl_rng * r) ;
